+++ src/fv_params.cpp
@@ -5,13 +5,13 @@
 
 using namespace rlwe::fv;
         
-KeyParameters::KeyParameters(uint32_t n, uint32_t q, uint32_t t) : 
+KeyParameters::KeyParameters(size_t n, uint32_t q, uint32_t t) : 
   KeyParameters(n, ZZ(q), ZZ(t)) {}
 
-KeyParameters::KeyParameters(uint32_t n, const ZZ & q, const ZZ & t) : 
+KeyParameters::KeyParameters(size_t n, const ZZ & q, const ZZ & t) : 
   KeyParameters(n, q, t, DEFAULT_DECOMPOSITION_BIT_COUNT, DEFAULT_ERROR_STANDARD_DEVIATION) {}
 
-KeyParameters::KeyParameters(uint32_t n, const ZZ & q, const ZZ & t, uint32_t log_w, float sigma) : 
+KeyParameters::KeyParameters(size_t n, const ZZ & q, const ZZ & t, uint32_t log_w, float sigma) : 
   n(n), q(q), t(t), log_w(log_w), sigma(sigma), delta(q / t) 
 {
   // Assert that n is even, assume that it is a power of 2
